Clean up Markdown duplicate
