+++ src/Aspire.Dashboard/Model/Assistant/ChatViewModelBuilder.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Aspire.Dashboard.Model.Assistant.Markdown;
+using Aspire.Dashboard.Model.Markdown;
 
 namespace Aspire.Dashboard.Model.Assistant;
