+++ tests/Aspire.Dashboard.Tests/Markdown/MarkdownProcessorTests.cs
@@ -2,8 +2,15 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Text.RegularExpressions;
+using Aspire.Dashboard.Model.Assistant;
+using Aspire.Dashboard.Model.Assistant.Markdown;
 using Aspire.Dashboard.Model.Markdown;
+using Aspire.Dashboard.Otlp.Storage;
 using Aspire.Dashboard.Resources;
+using Aspire.Dashboard.Tests.Integration.Playwright.Infrastructure;
+using Aspire.Tests.Shared.DashboardModel;
+using Aspire.Tests.Shared.Telemetry;
+using Markdig;
 using Xunit;
 
 namespace Aspire.Dashboard.Tests.Markdown;
@@ -330,8 +337,45 @@ public class MarkdownProcessorTests
         Assert.Equal(expected, html.Trim(), ignoreLineEndingDifferences: true);
     }
 
-    internal static MarkdownProcessor CreateMarkdownProcessor(HashSet<string>? safeUrlSchemes = null)
+    [Theory]
+    [InlineData("frontend")]
+    [InlineData("frontend-abcxyz")]
+    [InlineData("**frontend-abcxyz**")]
+    [InlineData("*_frontend-abcxyz_*")]
+    public void ToHtml_ResourceCode_ConvertedToLink(string code)
     {
-        return new MarkdownProcessor(new TestStringLocalizer<ControlsStrings>(), safeUrlSchemes);
+        // Arrange
+        var dashboardClient = new MockDashboardClient(resources: [ModelTestHelpers.CreateResource(resourceName: "frontend-abcxyz", displayName: "frontend")]);
+        var context = CreateAssistantChatDataContext(dashboardClient: dashboardClient);
+        var processor = CreateMarkdownProcessor(extensions: [new AspireEnrichmentExtension(new AspireEnrichmentOptions { DataContext = context })]);
+
+        var markdown =
+            $"""
+            Test: `{code}`
+            """;
+
+        // Act
+        var html = processor.ToHtml(markdown, inCompleteDocument: true);
+
+        // Assert
+        Assert.Contains("""
+            href="/?resource=frontend-abcxyz"
+            """, html.Trim());
+    }
+
+    internal static AssistantChatDataContext CreateAssistantChatDataContext(TelemetryRepository? telemetryRepository = null, IDashboardClient? dashboardClient = null)
+    {
+        var context = new AssistantChatDataContext(
+            telemetryRepository ?? TelemetryTestHelpers.CreateRepository(),
+            dashboardClient ?? new MockDashboardClient(),
+            [],
+            new TestStringLocalizer<Dashboard.Resources.AIAssistant>());
+
+        return context;
+    }
+
+    internal static MarkdownProcessor CreateMarkdownProcessor(HashSet<string>? safeUrlSchemes = null, List<IMarkdownExtension>? extensions = null)
+    {
+        return new MarkdownProcessor(new TestStringLocalizer<ControlsStrings>(), safeUrlSchemes, extensions ?? []);
     }
 }
