--- tests/Aspire.Dashboard.Tests/Model/AIAssistant/MarkdownProcessorTests.cs
@@ -1,356 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Text.RegularExpressions;
-using Aspire.Dashboard.Model.Assistant;
-using Aspire.Dashboard.Model.Assistant.Markdown;
-using Aspire.Dashboard.Otlp.Storage;
-using Aspire.Dashboard.Resources;
-using Aspire.Dashboard.Tests.Integration.Playwright.Infrastructure;
-using Aspire.Tests.Shared.DashboardModel;
-using Aspire.Tests.Shared.Telemetry;
-using Xunit;
-
-namespace Aspire.Dashboard.Tests.Model.AIAssistant;
-
-public class MarkdownProcessorTests
-{
-    [Fact]
-    public void ToHtml_FencedCodeBlockBreaksIndentation_TrailingTextIsntACodeBlock()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            To restart the unhealthy container resources effectively, follow these steps:
-
-            1. **Identify the Unhealthy Containers**:
-               - From the resource graph, note the names of the unhealthy containers (e.g., `postgres-euaaphdw`, `basketcache-urvebuku`, `messaging-ffc632b9`).
-
-            2. **Restart Containers Individually**:
-               - Use the appropriate container management tool (e.g., Docker CLI, Docker Compose, Kubernetes) to restart each container. For Docker, you can use:
-             ```bash
-
-             docker restart <container_name>
-
-             ```
-                 Replace `<container_name>` with the actual container name (e.g., `postgres-euaaphdw`).
-
-            3. **Verify Restart Success**
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        var count = Regex.Matches(html, Regex.Escape("code-block")).Count;
-        Assert.Equal(1, count); // There should be one code block. There shouldn't be an indented code block.
-    }
-
-    [Fact]
-    public void ToHtml_StartFencedCodeBlock_NoHtmlCode()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            Test:
-            ```csha
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            """
-            <p>Test:</p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Fact]
-    public void ToHtml_FencedCodeBlockEmpty_HasHtmlCode()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            Test:
-            ```csharp
-            ```
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Contains("</code>", html);
-    }
-
-    [Fact]
-    public void ToHtml_ContainsFencedCodeBlock_HtmlHasCode()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            Test:
-            ```csha
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            """
-            <p>Test:</p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Fact]
-    public void ToHtml_IncompleteFencedCodeBlock_HtmlHasCode()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            Test:
-            ```
-            In code bl...
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Contains("</code>", html);
-    }
-
-    [Fact]
-    public void ToHtml_ContainsFencedCodeBlockCsharp_HtmlHasCode()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            Test:
-            ```csharp
-            In code block.
-            ```
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Contains("language-csharp", html);
-        Assert.Contains("</code>", html);
-    }
-
-    [Fact]
-    public void ToHtml_ContainsHtml_HtmlIgnored()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            **Waiting Resources:** <b>test</b>
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            """
-            <p><strong>Waiting Resources:</strong> &lt;b&gt;test&lt;/b&gt;</p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Fact]
-    public void ToHtml_UrlWithJavaScript_UrlRemoved()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            """
-            [malicious](javascript:alert('hi'))
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            """
-            <p><a href="">malicious</a></p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Theory]
-    [InlineData("http://localhost:8080")]
-    [InlineData("https://localhost:8081")]
-    [InlineData("mailto:test@test.com")]
-    public void ToHtml_UrlWithValidSchemes_LinkAdded(string url)
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            $"""
-            [test]({url})
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            $"""
-            <p><a href="{url}" target="_blank" rel="noopener noreferrer nofollow">test</a></p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Theory]
-    [InlineData("http://contoso.com:8080", "http://contoso.com:8080")]
-    [InlineData("https://contoso.com:8081", "https://contoso.com:8081")]
-    [InlineData("mailto:test@test.com", "test@test.com")]
-    public void ToHtml_AutoLinkWithValidSchemes_LinkAdded(string url, string text)
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            $"""
-            {url}
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            $"""
-            <p><a href="{url}" target="_blank" rel="noopener noreferrer nofollow">{text}</a></p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Theory]
-    [InlineData("http://localhost:8080", "http://localhost:8080")]
-    [InlineData("https://localhost", "https://localhost")]
-    public void ToHtml_Localhost_LinkAdded(string url, string text)
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            $"""
-            {url}
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            $"""
-            <p><a href="{url}" target="_blank" rel="noopener noreferrer nofollow">{text}</a></p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Theory]
-    [InlineData("http://localhost:8080", "http://localhost:8080")]
-    [InlineData("https://localhost", "https://localhost")]
-    public void ToHtml_UrlInsideCode_LinkAdded(string url, string text)
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            $"""
-            `{url}`
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            $"""
-            <p><a href="{url}" target="_blank" rel="noopener noreferrer nofollow">{text}</a></p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Fact]
-    public void ToHtml_UrlInsideCodeWithExtra_NoLink()
-    {
-        // Arrange
-        var processor = CreateMarkdownProcessor();
-
-        var markdown =
-            $"""
-            `http://localhost:8080 extra`
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Equal(
-            $"""
-            <p><code>http://localhost:8080 extra</code></p>
-            """, html.Trim(), ignoreLineEndingDifferences: true);
-    }
-
-    [Theory]
-    [InlineData("frontend")]
-    [InlineData("frontend-abcxyz")]
-    [InlineData("**frontend-abcxyz**")]
-    [InlineData("*_frontend-abcxyz_*")]
-    public void ToHtml_ResourceCode_ConvertedToLink(string code)
-    {
-        // Arrange
-        var dashboardClient = new MockDashboardClient(resources: [ModelTestHelpers.CreateResource(resourceName: "frontend-abcxyz", displayName: "frontend")]);
-        var processor = CreateMarkdownProcessor(dashboardClient: dashboardClient);
-
-        var markdown =
-            $"""
-            Test: `{code}`
-            """;
-
-        // Act
-        var html = processor.ToHtml(markdown, inCompleteDocument: true);
-
-        // Assert
-        Assert.Contains("""
-            href="/?resource=frontend-abcxyz"
-            """, html.Trim());
-    }
-
-    internal static AssistantChatDataContext CreateAssistantChatDataContext(TelemetryRepository? telemetryRepository = null, IDashboardClient? dashboardClient = null)
-    {
-        var context = new AssistantChatDataContext(
-            telemetryRepository ?? TelemetryTestHelpers.CreateRepository(),
-            dashboardClient ?? new MockDashboardClient(),
-            [],
-            new TestStringLocalizer<Dashboard.Resources.AIAssistant>());
-
-        return context;
-    }
-
-    internal static MarkdownProcessor CreateMarkdownProcessor(TelemetryRepository? telemetryRepository = null, IDashboardClient? dashboardClient = null)
-    {
-        var context = CreateAssistantChatDataContext(telemetryRepository, dashboardClient);
-
-        return new MarkdownProcessor(context, new TestStringLocalizer<ControlsStrings>());
-    }
-}
